Added error logging, fixed backend transactions "created_at" column in database.py, added TODO, added extra link that can reach our website.  (#6)
* deleted "from backend" before import database, schemas

idk why its running for deepak, but it's not running on render either, which is a problem

* line 526 on database.py

* reversed line 526 in database.py

* Comment out created_at timestamp setting at line 334

* rolled back on commented out 334

* Print error logging for Supabase requests

Added error handling for Supabase response.

* ## TODO  move search bar to sticky on bottom on mobile view allow users to cancel orders if they regret buying, so that the buyer won't need to list all the details again. 


## TODO 
move search bar to sticky on bottom on mobile view
allow users to cancel orders if they regret buying, so that the buyer won't need to list all the details again.

* Moved TODO to lower section and added mobile support feature desc

Moved TODO section lower and added mobile support details.